Relocating of math class
