+++ MathLibrary/MathLibrary.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30320.27
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathLibrary", "MathLibrary\MathLibrary.csproj", "{CF1F772C-5DA2-4F45-9210-C6E8E97BB22D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathTest", "MathTest\MathTest.csproj", "{A38D3BF5-8F22-49A6-8D7E-EAE35EE20D32}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CF1F772C-5DA2-4F45-9210-C6E8E97BB22D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CF1F772C-5DA2-4F45-9210-C6E8E97BB22D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CF1F772C-5DA2-4F45-9210-C6E8E97BB22D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CF1F772C-5DA2-4F45-9210-C6E8E97BB22D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A38D3BF5-8F22-49A6-8D7E-EAE35EE20D32}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A38D3BF5-8F22-49A6-8D7E-EAE35EE20D32}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A38D3BF5-8F22-49A6-8D7E-EAE35EE20D32}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A38D3BF5-8F22-49A6-8D7E-EAE35EE20D32}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C70D38F9-DD7B-4EE5-AF4B-40927C390ED3}
+	EndGlobalSection
+EndGlobal
